Fix for Registration docs error, inheriting from Registry
The sample "FooBarRegistry" is shown as subclassing Registry and
not ServiceRegistry. This was due to the doc tools picking up the
sample from the old StructureMap testing code.

I've picked up and copied the relevant code into the
Lamar.Testing.Samples namespace and renamted the original
code samples in the StructureMap test code.