Add failing test proving ContainerValidationException for IServiceScopeFactory
When we run AssertConfigurationIsValid and there is a service that depends on IServiceScopeFactory, the AssertConfigurationIsValid throws ContainerValidationException while it should not.

The services.AddHealthChecks() configures DefaultHealthCheckService which depends on IServiceScopeFactory. These are all internal classes that we can't change to not use IServiceScopeFactory.

I didn't notice problems with using healthchecks in the api context - i.e. they work fine, DefaultHealthCheckService is resolved correctly and IServiceScopeFactory invokes CreateScope() correctly. That's why I think that this is a problem specific only to when running AssertConfigurationIsValid()